feat(cuda): Support float16

+++ cuda/benchmark.py
@@ -26,7 +26,7 @@ from attention import ATTENTION
 
 
 device = torch.device("cuda") if options.cuda else torch.device("cpu")
-dtype = torch.float64 if options.double else torch.float32
+dtype = torch.float64 if options.double else torch.float16
 
 kwargs = {'dtype': dtype,
           'device': device,
